app/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
APScheduler==3.11.0
argcomplete==3.6.2
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
boto3==1.38.19
botocore==1.38.19
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
commitizen==4.7.2
commitlint==1.3.0
coverage==7.8.0
cron-descriptor==1.4.5
cryptography==45.0.2
decli==0.6.2
decorator==5.2.1
distlib==0.3.9
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-smtp-ssl==1.0
django-ssl-smtp==2.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
ecdsa==0.18.0
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
flake8==7.2.0
Flake8-pyproject==1.2.3
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
identify==2.6.10
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
ipython==8.27.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
nodeenv==1.9.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pluralizer==1.2.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.0
redis==6.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.0
rsa==4.0
ruff==0.11.10
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
slack-sdk==3.35.0
termcolor==2.5.0
tomli==2.2.1
tomlkit==0.13.2
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
